upgrade klayout to 0.25.6
Summary: This is a rebuild of D4587. Since F.von Gellhorn's diff was accepted, i'm submitting a new diff.

Test Plan: As before, open a few GDS files (kB to GB in size) and browse the layout. Open Help->About and check it ruby/python bindings are present in the `Build options'.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake

Differential Revision: https://dev.getsol.us/D4924